requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
boto3==1.28.85
python-dotenv==1.0.0
Pillow
numpy
requests>=2.31.0